requirements.txt
amqp==5.2.0
anyio==4.3.0
APScheduler==3.6.3
asgiref==3.7.2
billiard==4.2.0
cachetools==4.2.2
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-timezone-field==6.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
kombu==5.3.5
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2024.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
